Push

+++ proyecto2/othello_blai/NegaMaxAB.h
@@ -6,7 +6,7 @@ class NegaMaxAB_ {
 
 public:
 
-    long nodosGenerados;
+    unsigned long long nodosGenerados;
     
     int NegaMaxAB(state_t state, int depth, bool color,int alpha,int beta) {
